[NET-526] Toast component is now located at module level, set default parameters, inserted missing docs and headers
