+++ app/src/app/layout/toast/toast.module.ts
@@ -1,9 +1,26 @@
+/*
+* DATAGERRY - OpenSource Enterprise CMDB
+* Copyright (C) 2020 NETHINKS GmbH
+*
+* This program is free software: you can redistribute it and/or modify
+* it under the terms of the GNU Affero General Public License as
+* published by the Free Software Foundation, either version 3 of the
+* License, or (at your option) any later version.
+*
+* This program is distributed in the hope that it will be useful,
+* but WITHOUT ANY WARRANTY; without even the implied warranty of
+* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+* GNU Affero General Public License for more details.
+
+* You should have received a copy of the GNU Affero General Public License
+* along with this program.  If not, see <https://www.gnu.org/licenses/>.
+*/
+
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ToastContainerComponent } from './toast-container/toast-container.component';
 import { NgbToastModule } from '@ng-bootstrap/ng-bootstrap';
-import { ToastComponentComponent } from './toast-container/toast-component/toast-component.component';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { ToastComponentComponent } from './toast-component/toast-component.component';
 
 @NgModule({
   declarations: [ToastContainerComponent, ToastComponentComponent],
